third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_word_19.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'word'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} word`
  );
}

function makeSample(offset){
  const kSample = '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right word
testIt('ltr', 0, 'right', '^\u05D6|;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 1, 'right', '\u05D6^;\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 2, 'right', '\u05D6;^\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 3, 'right', '\u05D6;\u05D7^;\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 4, 'right', '\u05D6;\u05D7;^\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 5, 'right', '\u05D6;\u05D7;\u05D8^; \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 6, 'right', '\u05D6;\u05D7;\u05D8;^ \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 7, 'right', '\u05D6;\u05D7;\u05D8; ^\u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 8, 'right', '\u05D6;\u05D7;\u05D8; \u1497^;\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 9, 'right', '\u05D6;\u05D7;\u05D8; \u1497;^\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 10, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498^;\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 11, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;^\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 12, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499^; \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 13, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499;^ \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 14, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; ^\u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 15, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500^;\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 16, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;^\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 17, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501^;\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 18, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;^\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 19, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502^; '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 20, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502;^ '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 21, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; ^'\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 22, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '^\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 23, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A^;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 24, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;^they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 25, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;t^hey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 26, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;th^ey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 27, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;the^y| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 28, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they^ said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 29, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they ^said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 30, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they s^aid| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 31, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sa^id| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 32, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sai^d| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 33, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said^ "\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 34, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said ^"\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 35, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "^\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 36, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B^;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 37, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;^car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 38, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;c^ar| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 39, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 40, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 41, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 42, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 43, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 44, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 45, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 46, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 47, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 48, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 49, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 50, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 51, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\u202C;"\u202C;'?');
testIt('ltr', 52, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\u202C;"\u202C;'?');
testIt('ltr', 53, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;"\u202C;'?|');
testIt('ltr', 54, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;"\u202C;'?|');
testIt('ltr', 55, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;"\u202C;'?|');
testIt('ltr', 56, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^"\u202C;'?|');
testIt('ltr', 57, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"^\u202C;'?|');
testIt('ltr', 58, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C^;'?|');
testIt('ltr', 59, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;^'?|');
testIt('ltr', 60, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'^?|');

// ltr left word
testIt('ltr', 0, 'left', '|\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 1, 'left', '|\u05D6^;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 2, 'left', '|\u05D6;^\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 3, 'left', '\u05D6;|\u05D7^;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 4, 'left', '\u05D6;|\u05D7;^\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 5, 'left', '\u05D6;\u05D7;|\u05D8^; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 6, 'left', '\u05D6;\u05D7;|\u05D8;^ \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 7, 'left', '\u05D6;\u05D7;|\u05D8; ^\u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 8, 'left', '\u05D6;\u05D7;\u05D8; |\u1497^;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 9, 'left', '\u05D6;\u05D7;\u05D8; |\u1497;^\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 10, 'left', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498^;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 11, 'left', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498;^\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 12, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499^; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 13, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499;^ \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 14, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499; ^\u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 15, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500^;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 16, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500;^\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 17, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501^;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 18, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501;^\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 19, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502^; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 20, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502;^ '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 21, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; ^'\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 22, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '^\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 23, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A^;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 24, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A;^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 25, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 26, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 27, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 28, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 29, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 30, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 31, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 32, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 33, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 34, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 35, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 36, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 37, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 38, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 39, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 40, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 41, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 42, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 43, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 44, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 45, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 46, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 47, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 48, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 49, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 50, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 51, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 52, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 53, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\u202C;"\u202C;'?');
testIt('ltr', 54, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\u202C;"\u202C;'?');
testIt('ltr', 55, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C^;"\u202C;'?');
testIt('ltr', 56, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;^"\u202C;'?');
testIt('ltr', 57, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"^\u202C;'?');
testIt('ltr', 58, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C^;'?');
testIt('ltr', 59, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;^'?');
testIt('ltr', 60, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;'^?');

// ltr forward word
testIt('ltr', 0, 'forward', '^\u05D6|;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 1, 'forward', '\u05D6^;\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 2, 'forward', '\u05D6;^\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 3, 'forward', '\u05D6;\u05D7^;\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 4, 'forward', '\u05D6;\u05D7;^\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 5, 'forward', '\u05D6;\u05D7;\u05D8^; \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 6, 'forward', '\u05D6;\u05D7;\u05D8;^ \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 7, 'forward', '\u05D6;\u05D7;\u05D8; ^\u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 8, 'forward', '\u05D6;\u05D7;\u05D8; \u1497^;\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 9, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;^\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 10, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498^;\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 11, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;^\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 12, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499^; \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 13, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499;^ \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 14, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; ^\u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 15, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500^;\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 16, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;^\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 17, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501^;\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 18, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;^\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 19, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502^; '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 20, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502;^ '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 21, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; ^'\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 22, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '^\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 23, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A^;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 24, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;^they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 25, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;t^hey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 26, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;th^ey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 27, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;the^y| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 28, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they^ said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 29, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they ^said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 30, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they s^aid| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 31, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sa^id| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 32, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sai^d| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 33, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said^ "\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 34, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said ^"\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 35, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "^\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 36, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B^;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 37, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;^car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 38, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;c^ar| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 39, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 40, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 41, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 42, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 43, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 44, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 45, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 46, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 47, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 48, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 49, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 50, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 51, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\u202C;"\u202C;'?');
testIt('ltr', 52, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\u202C;"\u202C;'?');
testIt('ltr', 53, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;"\u202C;'?|');
testIt('ltr', 54, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;"\u202C;'?|');
testIt('ltr', 55, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;"\u202C;'?|');
testIt('ltr', 56, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^"\u202C;'?|');
testIt('ltr', 57, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"^\u202C;'?|');
testIt('ltr', 58, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C^;'?|');
testIt('ltr', 59, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;^'?|');
testIt('ltr', 60, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'^?|');

// ltr backward word
testIt('ltr', 0, 'backward', '|\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 1, 'backward', '|\u05D6^;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 2, 'backward', '|\u05D6;^\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 3, 'backward', '\u05D6;|\u05D7^;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 4, 'backward', '\u05D6;|\u05D7;^\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 5, 'backward', '\u05D6;\u05D7;|\u05D8^; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 6, 'backward', '\u05D6;\u05D7;|\u05D8;^ \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 7, 'backward', '\u05D6;\u05D7;|\u05D8; ^\u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 8, 'backward', '\u05D6;\u05D7;\u05D8; |\u1497^;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 9, 'backward', '\u05D6;\u05D7;\u05D8; |\u1497;^\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 10, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498^;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 11, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498;^\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 12, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499^; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 13, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499;^ \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 14, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499; ^\u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 15, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500^;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 16, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500;^\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 17, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501^;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 18, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501;^\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 19, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502^; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 20, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502;^ '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 21, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; ^'\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 22, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '^\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 23, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A^;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 24, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A;^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 25, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 26, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 27, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 28, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 29, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 30, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 31, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 32, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 33, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 34, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 35, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 36, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 37, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 38, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 39, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 40, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 41, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 42, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 43, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 44, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 45, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 46, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 47, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 48, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 49, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 50, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 51, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 52, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\u202C;"\u202C;'?');
testIt('ltr', 53, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\u202C;"\u202C;'?');
testIt('ltr', 54, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\u202C;"\u202C;'?');
testIt('ltr', 55, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C^;"\u202C;'?');
testIt('ltr', 56, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;^"\u202C;'?');
testIt('ltr', 57, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"^\u202C;'?');
testIt('ltr', 58, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C^;'?');
testIt('ltr', 59, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;^'?');
testIt('ltr', 60, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;'^?');

// rtl right word
testIt('rtl', 0, 'right', '^\u05D6|;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 1, 'right', '\u05D6^;\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 2, 'right', '\u05D6;^\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 3, 'right', '\u05D6;\u05D7^;\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 4, 'right', '\u05D6;\u05D7;^\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 5, 'right', '\u05D6;\u05D7;\u05D8^; \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 6, 'right', '\u05D6;\u05D7;\u05D8;^ \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 7, 'right', '\u05D6;\u05D7;\u05D8; ^\u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 8, 'right', '\u05D6;\u05D7;\u05D8; \u1497^;\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 9, 'right', '\u05D6;\u05D7;\u05D8; \u1497;^\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 10, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498^;\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 11, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;^\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 12, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499^; \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 13, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499;^ \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 14, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; ^\u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 15, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500^;\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 16, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;^\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 17, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501^;\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 18, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;^\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 19, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502^; '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 20, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502;^ '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 21, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; ^'\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 22, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '^\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 23, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A^;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 24, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;^they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 25, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;t^hey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 26, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;th^ey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 27, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;the^y| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 28, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they^ said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 29, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they ^said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 30, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they s^aid| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 31, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sa^id| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 32, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sai^d| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 33, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said^ "\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 34, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said ^"\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 35, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "^\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 36, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B^;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 37, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;^car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 38, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;c^ar| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 39, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 40, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 41, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 42, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 43, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 44, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 45, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 46, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 47, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 48, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 49, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 50, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 51, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\u202C;"\u202C;'?');
testIt('rtl', 52, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\u202C;"\u202C;'?');
testIt('rtl', 53, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;"\u202C;'?|');
testIt('rtl', 54, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;"\u202C;'?|');
testIt('rtl', 55, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;"\u202C;'?|');
testIt('rtl', 56, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^"\u202C;'?|');
testIt('rtl', 57, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"^\u202C;'?|');
testIt('rtl', 58, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C^;'?|');
testIt('rtl', 59, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;^'?|');
testIt('rtl', 60, 'right', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'^?|');

// rtl left word
testIt('rtl', 0, 'left', '|\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 1, 'left', '|\u05D6^;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 2, 'left', '|\u05D6;^\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 3, 'left', '\u05D6;|\u05D7^;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 4, 'left', '\u05D6;|\u05D7;^\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 5, 'left', '\u05D6;\u05D7;|\u05D8^; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 6, 'left', '\u05D6;\u05D7;|\u05D8;^ \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 7, 'left', '\u05D6;\u05D7;|\u05D8; ^\u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 8, 'left', '\u05D6;\u05D7;\u05D8; |\u1497^;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 9, 'left', '\u05D6;\u05D7;\u05D8; |\u1497;^\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 10, 'left', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498^;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 11, 'left', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498;^\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 12, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499^; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 13, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499;^ \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 14, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499; ^\u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 15, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500^;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 16, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500;^\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 17, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501^;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 18, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501;^\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 19, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502^; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 20, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502;^ '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 21, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; ^'\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 22, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '^\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 23, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A^;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 24, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A;^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 25, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 26, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 27, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 28, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 29, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 30, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 31, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 32, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 33, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 34, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 35, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 36, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 37, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 38, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 39, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 40, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 41, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 42, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 43, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 44, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 45, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 46, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 47, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 48, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 49, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 50, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 51, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 52, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 53, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\u202C;"\u202C;'?');
testIt('rtl', 54, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\u202C;"\u202C;'?');
testIt('rtl', 55, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C^;"\u202C;'?');
testIt('rtl', 56, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;^"\u202C;'?');
testIt('rtl', 57, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"^\u202C;'?');
testIt('rtl', 58, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C^;'?');
testIt('rtl', 59, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;^'?');
testIt('rtl', 60, 'left', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;'^?');

// rtl forward word
testIt('rtl', 0, 'forward', '^\u05D6|;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 1, 'forward', '\u05D6^;\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 2, 'forward', '\u05D6;^\u05D7|;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 3, 'forward', '\u05D6;\u05D7^;\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 4, 'forward', '\u05D6;\u05D7;^\u05D8|; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 5, 'forward', '\u05D6;\u05D7;\u05D8^; \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 6, 'forward', '\u05D6;\u05D7;\u05D8;^ \u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 7, 'forward', '\u05D6;\u05D7;\u05D8; ^\u1497|;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 8, 'forward', '\u05D6;\u05D7;\u05D8; \u1497^;\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 9, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;^\u1498|;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 10, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498^;\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 11, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;^\u1499|; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 12, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499^; \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 13, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499;^ \u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 14, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; ^\u1500|;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 15, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500^;\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 16, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;^\u1501|;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 17, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501^;\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 18, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;^\u1502|; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 19, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502^; '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 20, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502;^ '\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 21, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; ^'\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 22, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '^\u202A;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 23, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A^;they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 24, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;^they| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 25, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;t^hey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 26, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;th^ey| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 27, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;the^y| said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 28, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they^ said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 29, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they ^said| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 30, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they s^aid| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 31, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sa^id| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 32, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they sai^d| "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 33, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said^ "\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 34, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said ^"\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 35, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "^\u202B;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 36, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B^;car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 37, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;^car| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 38, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;c^ar| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 39, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;ca^r| \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 40, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car^ \u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 41, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car ^\u05D3|;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 42, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3^;\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 43, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;^\u05D4|;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 44, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4^;\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 45, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;^\u05D5|; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 46, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5^; \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 47, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5;^ \u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 48, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; ^\u05D0|;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 49, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0^;\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 50, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;^\u05D1|;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 51, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1^;\u05D2|;\u202C;"\u202C;'?');
testIt('rtl', 52, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;^\u05D2|;\u202C;"\u202C;'?');
testIt('rtl', 53, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2^;\u202C;"\u202C;'?|');
testIt('rtl', 54, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;^\u202C;"\u202C;'?|');
testIt('rtl', 55, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C^;"\u202C;'?|');
testIt('rtl', 56, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;^"\u202C;'?|');
testIt('rtl', 57, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"^\u202C;'?|');
testIt('rtl', 58, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C^;'?|');
testIt('rtl', 59, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;^'?|');
testIt('rtl', 60, 'forward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'^?|');

// rtl backward word
testIt('rtl', 0, 'backward', '|\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 1, 'backward', '|\u05D6^;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 2, 'backward', '|\u05D6;^\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 3, 'backward', '\u05D6;|\u05D7^;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 4, 'backward', '\u05D6;|\u05D7;^\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 5, 'backward', '\u05D6;\u05D7;|\u05D8^; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 6, 'backward', '\u05D6;\u05D7;|\u05D8;^ \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 7, 'backward', '\u05D6;\u05D7;|\u05D8; ^\u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 8, 'backward', '\u05D6;\u05D7;\u05D8; |\u1497^;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 9, 'backward', '\u05D6;\u05D7;\u05D8; |\u1497;^\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 10, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498^;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 11, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;|\u1498;^\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 12, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499^; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 13, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499;^ \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 14, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;|\u1499; ^\u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 15, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500^;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 16, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; |\u1500;^\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 17, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501^;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 18, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;|\u1501;^\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 19, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502^; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 20, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502;^ '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 21, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; ^'\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 22, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '^\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 23, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A^;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 24, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;|\u1502; '\u202A;^they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 25, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|t^hey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 26, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|th^ey said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 27, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|the^y said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 28, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they^ said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 29, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;|they ^said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 30, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |s^aid "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 31, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sa^id "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 32, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |sai^d "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 33, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said^ "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 34, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said ^"\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 35, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "^\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 36, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B^;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 37, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they |said "\u202B;^car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 38, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|c^ar \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 39, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|ca^r \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 40, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car^ \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 41, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;|car ^\u05D3;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 42, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3^;\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 43, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car |\u05D3;^\u05D4;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 44, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4^;\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 45, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;|\u05D4;^\u05D5; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 46, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5^; \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 47, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5;^ \u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 48, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;|\u05D5; ^\u05D0;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 49, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0^;\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 50, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; |\u05D0;^\u05D1;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 51, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1^;\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 52, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;|\u05D1;^\u05D2;\u202C;"\u202C;'?');
testIt('rtl', 53, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2^;\u202C;"\u202C;'?');
testIt('rtl', 54, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;^\u202C;"\u202C;'?');
testIt('rtl', 55, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C^;"\u202C;'?');
testIt('rtl', 56, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;^"\u202C;'?');
testIt('rtl', 57, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"^\u202C;'?');
testIt('rtl', 58, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C^;'?');
testIt('rtl', 59, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;^'?');
testIt('rtl', 60, 'backward', '\u05D6;\u05D7;\u05D8; \u1497;\u1498;\u1499; \u1500;\u1501;\u1502; '\u202A;they said "\u202B;car \u05D3;\u05D4;\u05D5; \u05D0;\u05D1;|\u05D2;\u202C;"\u202C;'^?');
</script>
